{{cookiecutter.repo_name}}/requirements/common.txt
# Latest Stable release of Django
Django==1.7.5

# Extensions
# -------------------------------------
Werkzeug==0.10.1
django-extensions==1.5.0
pytz==2014.10
requests==2.5.1

# Configuration
# -------------------------------------
django-configurations==0.8
django-sites==0.8
django-secure==1.0.1
dj-database-url==0.3.0
python-dotenv==0.1.0

# Models
# -------------------------------------
django-model-utils==2.2
django-uuid-upload-path==1.0.0
django-versatileimagefield==0.5.3

# Images
# -------------------------------------
Pillow==2.7.0

# For the persistance stores
# -------------------------------------
psycopg2==2.6

# Unicode slugification
# -------------------------------------
unicode-slugify==0.1.3
django-autoslug==1.7.2

# Django Rest Framework
# -------------------------------------
djangorestframework==3.0.5
Markdown>=2.3.1,<2.5   # to fix conflict with mkdocs

{% if cookiecutter.celery == 'y' %}
celery==3.1.17
{% endif %}
